+++ packages/orq-rc/src/models/operations/createidentity.ts
@@ -139,7 +139,7 @@ export const CreateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-13T10:23:19.161Z",
+    "2026-02-13T11:03:16.021Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {